Upload file lagi guys

+++ pertemuan10/index.php
@@ -31,7 +31,7 @@ $mahasiswa = query("SELECT * FROM mahasiswa");
         <td><?= $i; ?></td>
         <td>
             <a href="">ubah</a> |
-            <a href="hapus.php?id=<?= $row["id"]; ?>">hapus</a>
+            <a href="hapus.php?id=<?= $row["id"]; ?>" onclick="return confirm('Yakin?');">hapus</a>
         </td>
         <td><img src="img/<?= $row["gambar"]; ?>" width="50"></td>
         <td><?= $row["nim"]; ?></td>
